fix(clippy): drop unnecessary u64 casts in storage_dashboard
rust 1.94 clippy flags `stat.fragment_size() as u64` etc. as
unnecessary_cast (the statvfs fields are already u64 on Linux). Drop the
casts. Blocking CI on PR #95 (CO-279) and PR #96 (CO-280).

+++ co-web/src/admin/storage_dashboard.rs
@@ -118,10 +118,10 @@ fn file_size(path: &Path) -> u64 {
 fn disk_stats(path: &Path) -> (u64, u64) {
     match nix::sys::statvfs::statvfs(path) {
         Ok(stat) => {
-            let frsize = stat.fragment_size() as u64;
+            let frsize = stat.fragment_size();
             (
-                stat.blocks() as u64 * frsize,
-                stat.blocks_available() as u64 * frsize,
+                stat.blocks() * frsize,
+                stat.blocks_available() * frsize,
             )
         }
         Err(e) => {
